twelfth iteration

+++ ArduinoTris.ino
@@ -23,9 +23,9 @@ const uint8_t boardHeight = 20;
 const uint8_t pulse = 2;
 const uint8_t latch = 3;
 const uint8_t data = 4;
-const uint8_t NUM_PIECES_PER_TETRAMINO = 4;
-const uint8_t X_INDEX = 0;
-const uint8_t Y_INDEX = 1;
+const int NUM_PIECES_PER_TETRAMINO = 4;
+const int X_INDEX = 0;
+const int Y_INDEX = 1;
 
 Controller controller(latch, pulse, data);
 
@@ -50,10 +50,10 @@ const int tetraminoPiecePositions[7][4][3][2] = {
   {{{-1,0},{1,0},{1,1}}, {{-1,1},{0,1},{0,-1}}, {{-1,-1},{-1,0},{1,0}}, {{0,1},{0,-1},{1,-1}}}
 };
 
-int8_t currPosX;
-int8_t currPosY;
-uint8_t currTetramino;
-uint8_t currRotation;
+int currPosX;
+int currPosY;
+int currTetramino;
+int currRotation;
 
 void setup() {
   // setup LEDs
@@ -89,8 +89,8 @@ void newGame() {
 }
 
 void clearBoard() {
-  for (uint8_t i = 0; i < boardWidth; i++) {
-    for (uint8_t j = 0; j < boardHeight; j++) {
+  for (int i = 0; i < boardWidth; i++) {
+    for (int j = 0; j < boardHeight; j++) {
       board[i][j] = 0;
     }
   }
@@ -115,9 +115,9 @@ void putTetraminoOnBoard() {
 
     // each tetramino consists of 4 pieces
     // loop through each piece and find its relative x & y position
-   for(uint8_t tetraminoPiece=0;tetraminoPiece<NUM_PIECES_PER_TETRAMINO;tetraminoPiece++){
-      int8_t pieceXPos = 0;
-      int8_t pieceYPos = 0;
+   for(int tetraminoPiece=0;tetraminoPiece<NUM_PIECES_PER_TETRAMINO;tetraminoPiece++){
+      int pieceXPos = 0;
+      int pieceYPos = 0;
 
       // always uses position (0,0) for last piece (saves memory space)
       // need to calculate the other pieces
@@ -127,8 +127,8 @@ void putTetraminoOnBoard() {
       }
 
       // add relative position to absolute position
-      int8_t xPos = currPosX + pieceXPos;
-      int8_t yPos = currPosY + pieceYPos;
+      int xPos = currPosX + pieceXPos;
+      int yPos = currPosY + pieceYPos;
 
       bool xInBounds = xPos >= 0 && xPos < boardWidth;
       bool yInBounds = yPos >= 0 && yPos < boardHeight;
@@ -141,9 +141,9 @@ void putTetraminoOnBoard() {
 }
 
 void clearFilledRows() {
-  for(uint8_t y = 0; y < boardHeight; y++) {
-    uint8_t numFilled = 0;
-    for(uint8_t x = 0; x < boardWidth; x++) {
+  for(int y = 0; y < boardHeight; y++) {
+    int numFilled = 0;
+    for(int x = 0; x < boardWidth; x++) {
       if(board[x][y] != 0) {
         numFilled++;
       }
@@ -151,13 +151,13 @@ void clearFilledRows() {
 
     if(numFilled == boardWidth) {
       // clear my row
-      for(uint8_t x = 0; x < boardWidth; x++) {
+      for(int x = 0; x < boardWidth; x++) {
         board[x][y] = 0;
       }
 
       // pull down all rows above me
-      for(uint8_t y2 = y-1; y2 >=0; y2--) {
-        for(uint8_t x = 0; x < boardWidth; x++) {
+      for(int y2 = y-1; y2 >=0; y2--) {
+        for(int x = 0; x < boardWidth; x++) {
           board[x][y2+1] = board[x][y2];
         }
       }
@@ -165,7 +165,7 @@ void clearFilledRows() {
   }
 }
 
-void moveIfClear(int8_t posX, int8_t posY, uint8_t rotation) {
+void moveIfClear(int posX, int posY, int rotation) {
   if(!isBlocked(posX, posY, rotation)) {
     currPosX = posX;
     currPosY = posY;
@@ -173,15 +173,15 @@ void moveIfClear(int8_t posX, int8_t posY, uint8_t rotation) {
   }
 }
 
-bool isBlocked(int8_t posX, int8_t posY, uint8_t rotation) {  
+bool isBlocked(int posX, int posY, int rotation) {  
  
-     const uint8_t LAST_PIECE = NUM_PIECES_PER_TETRAMINO - 1;
+     const int LAST_PIECE = NUM_PIECES_PER_TETRAMINO - 1;
 
     // each tetramino consists of 4 pieces
     // loop through each piece and find its relative x & y position
-   for(uint8_t tetraminoPiece=0;tetraminoPiece<NUM_PIECES_PER_TETRAMINO;tetraminoPiece++){
-      int8_t pieceXPos = 0;
-      int8_t pieceYPos = 0;
+   for(int tetraminoPiece=0;tetraminoPiece<NUM_PIECES_PER_TETRAMINO;tetraminoPiece++){
+      int pieceXPos = 0;
+      int pieceYPos = 0;
 
       // always uses position (0,0) for last piece (saves memory space)
       // need to calculate the other pieces
@@ -191,8 +191,8 @@ bool isBlocked(int8_t posX, int8_t posY, uint8_t rotation) {
       }
 
       // add relative position to absolute position
-      int8_t xPos = posX + pieceXPos;
-      int8_t yPos = posY + pieceYPos;
+      int xPos = posX + pieceXPos;
+      int yPos = posY + pieceYPos;
 
       bool xInBounds = xPos >= 0 && xPos < boardWidth;
       if(!xInBounds) {
@@ -211,13 +211,13 @@ bool isBlocked(int8_t posX, int8_t posY, uint8_t rotation) {
 
 void drawTetramino() {
   
-    const uint8_t LAST_PIECE = NUM_PIECES_PER_TETRAMINO - 1;
+    const int LAST_PIECE = NUM_PIECES_PER_TETRAMINO - 1;
 
     // each tetramino consists of 4 pieces
     // loop through each piece and find its relative x & y position
-   for(uint8_t tetraminoPiece=0;tetraminoPiece<NUM_PIECES_PER_TETRAMINO;tetraminoPiece++){
-      int8_t pieceXPos = 0;
-      int8_t pieceYPos = 0;
+   for(int tetraminoPiece=0;tetraminoPiece<NUM_PIECES_PER_TETRAMINO;tetraminoPiece++){
+      int pieceXPos = 0;
+      int pieceYPos = 0;
 
       // always uses position (0,0) for last piece (saves memory space)
       // need to calculate the other pieces
@@ -227,8 +227,8 @@ void drawTetramino() {
       }
 
       // add relative position to absolute position
-      int8_t xPos = currPosX + pieceXPos;
-      int8_t yPos = currPosY + pieceYPos;
+      int xPos = currPosX + pieceXPos;
+      int yPos = currPosY + pieceYPos;
 
       bool xInBounds = xPos >= 0 && xPos < boardWidth;
       bool yInBounds = yPos >= 0 && yPos < boardHeight;
@@ -241,8 +241,8 @@ void drawTetramino() {
 }
 
 void drawBoard() {
-  for(int8_t x = 0; x < boardWidth; x++) {
-    for (int8_t y = 0; y < boardHeight; y++) {
+  for(int x = 0; x < boardWidth; x++) {
+    for (int y = 0; y < boardHeight; y++) {
       strip.setPixelColor(x,y,board[x][y]);
     }
   }
@@ -262,7 +262,7 @@ void moveRight() {
 }
 
 void moveDown() {
-  uint8_t newPosY = currPosY+1;
+  int newPosY = currPosY+1;
   if(!isBlocked(currPosX, newPosY, currRotation)) {
       currPosY = newPosY;
   } else {
@@ -274,7 +274,7 @@ void moveDown() {
 }
 
 void spinClockwise() {
-  uint8_t newRotation = currRotation+1;
+  int newRotation = currRotation+1;
   if(newRotation > 3) {
     newRotation = 0;
   }
@@ -282,7 +282,7 @@ void spinClockwise() {
 }
 
 void spinCounterClockwise() {
-  int8_t newRotation = currRotation-1;
+  int newRotation = currRotation-1;
   if(newRotation < 0) {
     newRotation = 3;
   }
@@ -319,17 +319,17 @@ void handleInput() {
 
 void updateController() {
   // update last frame button info
-  for(uint8_t i = 0; i < 8; i++) {
+  for(int i = 0; i < 8; i++) {
     lastButtonsPressed[i] = currentButtonsPressed[i];
   }
   controller.latch();
   // update current frame button info
-  for(uint8_t i = 0; i < 8; i++) {
+  for(int i = 0; i < 8; i++) {
     currentButtonsPressed[i] = controller.pressed(i);
   }
 }
 
-bool isControllerTriggered(uint8_t button) {
+bool isControllerTriggered(int button) {
   return(currentButtonsPressed[button] && !lastButtonsPressed[button]);
 }
